requirements.txt
# ── LLM / Orchestration ───────────────────────────────────────────────────────
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
groq>=0.11.0

# ── Vector store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests>=2.32.0

# ── Numerics ──────────────────────────────────────────────────────────────────
numpy>=1.26.0

# ── Config / Secrets ──────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Logging ───────────────────────────────────────────────────────────────────
structlog>=24.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
